[2908] fixed svg, change text params
