+++ src/components/modular/WalletCard/index.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
-import { StyleSheet, View } from 'react-native';
+import { View } from 'react-native';
 import { useTranslation } from 'react-i18next';
 import { Row, Spacer, Spinner, Text } from '@components/base';
 import { CopyToClipboardButton } from '@components/composite';
 import { COLORS } from '@constants/colors';
-import { shadow } from '@constants/shadow';
 import { NumberUtils } from '@utils/number';
-import { moderateScale, scale, verticalScale } from '@utils/scaling';
+import { scale, verticalScale } from '@utils/scaling';
 import { StringUtils } from '@utils/string';
 import { LogoGradient } from '@components/svg/icons';
 import { ToastPosition } from '../Toast';
+import { styles } from './styles';
 
 export const WalletCardHeight = 172;
 export interface WalletCardProps {
@@ -44,25 +44,32 @@ export const WalletCard = (props: WalletCardProps) => {
       <View style={styles.logo}>
         <LogoGradient />
       </View>
-      <CopyToClipboardButton
-        textToDisplay={StringUtils.formatAddress(
-          address,
-          addressLeftPadding,
-          addressRightPadding
-        )}
-        textToCopy={address}
-        textProps={{
-          fontFamily: 'Inter_400Regular',
-          fontSize: 14,
-          fontWeight: '400',
-          color: addressTextColor
-        }}
-        iconProps={{
-          color: addressTextColor
-        }}
-        successTextProps={{ color: COLORS.neutral0, fontSize: 14 }}
-        toastProps={{ position: ToastPosition.Top }}
-      />
+        <CopyToClipboardButton
+          textToDisplay={StringUtils.formatAddress(
+            address,
+            addressLeftPadding,
+            addressRightPadding
+          )}
+          copiedTextWrapperStyle={{
+            backgroundColor: COLORS.lightWhite,
+            borderColor: COLORS.transparentWhite,
+            borderWidth: 1,
+            paddingHorizontal: 8,
+            borderRadius: 20
+          }}
+          textToCopy={address}
+          textProps={{
+            fontFamily: 'Inter_400Regular',
+            fontSize: 14,
+            fontWeight: '400',
+            color: addressTextColor
+          }}
+          iconProps={{
+            color: addressTextColor
+          }}
+          successTextProps={{ color: COLORS.neutral0, fontSize: 14 }}
+          toastProps={{ position: ToastPosition.Top }}
+        />
       <View>
         {balanceLoading ? (
           <Spinner />
@@ -110,30 +117,3 @@ export const WalletCard = (props: WalletCardProps) => {
     </View>
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    borderRadius: moderateScale(16),
-    // minHeight: 148,
-    width: scale(300),
-    height: verticalScale(172),
-    minHeight: WalletCardHeight,
-    overflow: 'hidden',
-    paddingLeft: scale(20),
-    paddingVertical: verticalScale(24),
-    justifyContent: 'space-between',
-    ...shadow
-  },
-  logo: {
-    position: 'absolute',
-    top: -verticalScale(16),
-    right: -scale(16)
-  },
-  usdPriceBg: {
-    justifyContent: 'center',
-    minHeight: verticalScale(20),
-    borderRadius: 1000,
-    backgroundColor: COLORS.alphaWhite5,
-    paddingHorizontal: scale(8)
-  }
-});
