+++ src/components/svg/icons/AddressTracking.tsx
@@ -35,14 +35,7 @@ export function AddressTrackingIcon(props: Omit<IconProps, 'color'>) {
         stroke="#0E0E0E"
         strokeWidth={2}
       />
-      <Mask
-        id="prefix__a"
-        maskUnits="userSpaceOnUse"
-        x={113}
-        y={120}
-        width={194}
-        height={82}
-      >
+      <Mask id="prefix__a" maskUnits="userSpaceOnUse">
         <Path
           d="M305.059 160.843c0 5.262-2.53 10.363-7.288 15.098-4.762 4.738-11.704 9.051-20.362 12.695-17.311 7.285-41.296 11.815-67.841 11.815-26.544 0-50.529-4.53-67.84-11.815-8.658-3.644-15.6-7.957-20.362-12.695-4.758-4.735-7.288-9.836-7.288-15.098 0-5.262 2.53-10.363 7.288-15.097 4.762-4.739 11.704-9.052 20.362-12.695 17.311-7.286 41.296-11.816 67.84-11.816 26.545 0 50.53 4.53 67.841 11.816 8.658 3.643 15.6 7.956 20.362 12.695 4.758 4.734 7.288 9.835 7.288 15.097z"
           fill="#fff"
